+++ src/minecraft/com/vbitz/MinecraftScript/MinecraftScriptMod.java
@@ -46,7 +46,7 @@ import cpw.mods.fml.common.registry.LanguageRegistry;
 import cpw.mods.fml.common.registry.TickRegistry;
 import cpw.mods.fml.relauncher.Side;
 
-@Mod(modid="MinecraftScript", name="MinecraftScript", version="1.2.1") // mental note, update this loads
+@Mod(modid="MinecraftScript", name="MinecraftScript", version="1.4.0") // mental note, update this loads
 @NetworkMod(clientSideRequired=false, serverSideRequired=true)
 public class MinecraftScriptMod {
 	@Instance("MinecraftScriptMod")